Rename files BaseManageOfferOpFrame to ManageOfferOpFrameBase, ManageOfferOpFrame to ManageSellOfferOpFrame, and CreatePassiveOfferOpFrame to CreatePassiveSellOfferOpFrame
