+++ src/transactions/ManageSellOfferOpFrame.h
@@ -4,7 +4,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "transactions/BaseManageOfferOpFrame.h"
+#include "transactions/ManageOfferOpFrameBase.h"
 
 namespace stellar
 {
